Update project file
