+++ src/Couchbase.EntityFrameworkCore/Couchbase.EntityFrameworkCore.csproj
@@ -5,19 +5,19 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
         <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-        <Version>1.0.0-RC.2</Version>
+        <Version>2.0.0-pre.1</Version>
         <Title>EF Core Couchbase DB Provider</Title>
         <Authors>Couchbase, Inc.</Authors>
-        <Copyright>Couchbase, Inc. (c) 2025</Copyright>
+        <Copyright>Couchbase, Inc. (c) 2026</Copyright>
         <PackageProjectUrl>https://github.com/couchbaselabs/couchbase-efcore-provider</PackageProjectUrl>
         <PackageLicenseUrl>https://github.com/couchbaselabs/couchbase-efcore-provider/blob/main/README.md</PackageLicenseUrl>
         <RepositoryUrl>https://github.com/couchbaselabs/couchbase-efcore-provider</RepositoryUrl>
-        <PackageTags>couchbase, json, efcore, entity, framework, ef, database, nosql</PackageTags>
+        <PackageTags>couchbase, json, efcore, entity, framework, ef, database, nosql, net10</PackageTags>
         <Description>An EF Core DB Provider for Couchbase NoSQL database written in CSharp</Description>
         <PackageIcon>couchbase.png</PackageIcon>
         <AssemblyOriginatorKeyFile>Couchbase.snk</AssemblyOriginatorKeyFile>
         <PackageReadmeFile>README.md</PackageReadmeFile>
-        <IsPackable>false</IsPackable>
+        <IsPackable>true</IsPackable>
     </PropertyGroup>
 
     <ItemGroup>
