--- tests/Couchbase.EntityFrameworkCore.UnitTests/CouchbaseDbDataReaderTests.cs
@@ -1,107 +1,0 @@
-using Couchbase.EntityFrameworkCore.Storage.Internal;
-using Couchbase.Query;
-using Moq;
-using Xunit;
-
-namespace Couchbase.EntityFrameworkCore.UnitTests
-{
-    public class CouchbaseDbDataReaderTest
-    {
-        [Fact]
-        public async Task Read_ReturnsTrue_WhenMoreRowsExist()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            var mockEnumerator = new Mock<IAsyncEnumerator<object>>();
-            mockEnumerator.SetupSequence(e => e.MoveNextAsync())
-                .Returns(new ValueTask<bool>(true))
-                .Returns(new ValueTask<bool>(false));
-            mockQueryResult.Setup(q => q.GetAsyncEnumerator(It.IsAny<CancellationToken>())).Returns(mockEnumerator.Object);
-
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act
-            var result = reader.Read();
-
-            // Assert
-            Assert.True(result);
-        }
-
-        [Fact]
-        public async Task Read_ReturnsFalse_WhenNoMoreRowsExist()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            var mockEnumerator = new Mock<IAsyncEnumerator<object>>();
-            mockEnumerator.Setup(e => e.MoveNextAsync()).Returns(new ValueTask<bool>(false));
-            mockQueryResult.Setup(q => q.GetAsyncEnumerator(It.IsAny<CancellationToken>())).Returns(mockEnumerator.Object);
-
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act
-            var result = reader.Read();
-
-            // Assert
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void HasRows_ReturnsTrue_WhenQueryStatusIsSuccess()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            mockQueryResult.Setup(q => q.MetaData.Status).Returns(QueryStatus.Success);
-
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act
-            var result = reader.HasRows;
-
-            // Assert
-            Assert.True(result);
-        }
-
-        [Fact]
-        public void HasRows_ReturnsFalse_WhenQueryStatusIsNotSuccess()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            mockQueryResult.Setup(q => q.MetaData.Status).Returns(QueryStatus.Errors);
-
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act
-            var result = reader.HasRows;
-
-            // Assert
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void IsClosed_ReturnsFalse_ByDefault()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act
-            var result = reader.IsClosed;
-
-            // Assert
-            Assert.False(result);
-        }
-
-        [Fact]
-        public void GetBoolean_ThrowsNotImplementedException()
-        {
-            // Arrange
-            var mockQueryResult = new Mock<IQueryResult<object>>();
-            var reader = new CouchbaseDbDataReader<object>(mockQueryResult.Object);
-
-            // Act & Assert
-            Assert.Throws<NotImplementedException>(() => reader.GetBoolean(0));
-        }
-
-        // Add similar tests for other Get methods (GetByte, GetChar, GetDateTime, etc.)
-    }
-}
